Move MojarraAfterBeanDiscoveryDecorator out of inner class
